+++ CHANGELOG.md
@@ -1,5 +1,41 @@
 # Change History
 
+## March 4 2020: v2.8.1
+
+  Minor fix and improvements.
+
+  * **Improvements**
+
+    - Uses a `sync.Pool` to preserve the connection buffers on close to remove pressure from the allocator and the GC during connection churns.
+
+  * **Fixes**
+
+    - Cleanup the data structure cross refs on Cluster.Close to help GC free the objects.
+
+## February 28 2020: v2.8.0
+
+  Minor feature release.
+
+  * **New Features**
+
+    - Allows `,omitempty` tag to be used in struct tags. It behaves the same as the stdlib json. Note that there should be no whitespace between the comma and the rest of the tag.
+
+## January 30 2020: v2.7.2
+
+  Minor fix release.
+
+  * **Fixes**
+
+    - Resolves an issue where an invalid/malformed compressed header could cause a panic by reporting wrong compressed data size. Resolves #280.
+
+## January 20 2020: v2.7.1
+
+  Minor fix release.
+
+  * **Fixes**
+
+    - Fixes an issue where an error was not checked after a read in `multi_command.go` and would cause a panic. Resolves #280.
+
 ## December 24 2019: v2.7.0
 
   Minor feature and fix release.